Extract review/persona handlers from mcp-handlers.ts into mcp-review-handlers.ts
mcp-handlers.ts was 1075 lines with all MCP handler implementations in a single
createMcpHandlers() function. Extracted the 8 review/persona-related handlers
(submitResolution, cancelRecheck, updateReviewStatus, completeReview,
getParentContext, getRecheckContext, listPersonas, launchPersona) into a new
createReviewHandlers() module. mcp-handlers.ts drops to 652 lines. The API
surface is unchanged — review handlers are spread back into the return object.
